feat(feature): update my-info update feature - my-info update feature 내 viewmodel 데이터들을 state로 분리합니다. - 개별 state로 분리함에 따른 관련문서를 작성하고, 해당 state들을 viewmodel과 연결합니다.
